- Removed vdi bindings in favour of those in ovdisis - Fixed the vid call tunneling to pass just enough parameters

+++ src/server/srv_global.h
@@ -7,7 +7,7 @@
 #define	STKSIZE	4096
 
 typedef struct globalvars {
-  WORD    vid;	      /* physical vdi id of the screen */
+  int     vid;	      /* physical vdi id of the screen */
   WORD    vmode;
   WORD    vmodecode;
   LONG    video;
